moved network record getters
