+++ pkg/internal/netolly/export/otel/metrics_test.go
@@ -10,7 +10,6 @@ import (
 	"github.com/grafana/beyla/pkg/internal/export/metric/attr"
 	"github.com/grafana/beyla/pkg/internal/export/otel"
 	"github.com/grafana/beyla/pkg/internal/netolly/ebpf"
-	"github.com/grafana/beyla/pkg/internal/netolly/export"
 )
 
 func TestMetricAttributes(t *testing.T) {
@@ -37,7 +36,7 @@ func TestMetricAttributes(t *testing.T) {
 	in.Id.DstIp.In6U.U6Addr8 = [16]uint8{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 33, 22, 11, 1}
 
 	me := &metricsExporter{metrics: &Expirer{attrs: metric.OpenTelemetryGetters(
-		export.NamedGetters, []attr.Name{
+		ebpf.RecordGetters, []attr.Name{
 			attr.SrcAddress, attr.DstAddres, attr.SrcPort, attr.DstPort, attr.SrcName, attr.DstName,
 			attr.K8sSrcName, attr.K8sSrcNamespace, attr.K8sDstName, attr.K8sDstNamespace,
 		})}}
@@ -85,7 +84,7 @@ func TestMetricAttributes_Filter(t *testing.T) {
 	in.Id.SrcIp.In6U.U6Addr8 = [16]uint8{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 12, 34, 56, 78}
 	in.Id.DstIp.In6U.U6Addr8 = [16]uint8{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 33, 22, 11, 1}
 
-	me := &Expirer{attrs: metric.OpenTelemetryGetters(export.NamedGetters, []attr.Name{
+	me := &Expirer{attrs: metric.OpenTelemetryGetters(ebpf.RecordGetters, []attr.Name{
 		"src.address",
 		"k8s.src.name",
 		"k8s.dst.name",
